// src/models/IoK8sApiAdmissionregistrationV1WebhookClientConfig/types.ts
import { IoK8sApiAdmissionregistrationV1ServiceReferenceModelOptions as IoK8sApiAdmissionregistrationV1ServiceReference } from '../IoK8sApiAdmissionregistrationV1ServiceReference';

export interface IoK8sApiAdmissionregistrationV1WebhookClientConfigModelOptions {
  caBundle?: string;
  service?: IoK8sApiAdmissionregistrationV1ServiceReference;
  url?: string;
}